refactor(core): deprecate optimized_container and rename typed_container to simd_batch
- Add [[deprecated]] attribute to optimized_container class
  with clear migration path to fast_policy_container
- Rename core::typed_container to core::simd_batch for clarity
- Create simd_batch.h with new class name and deprecated alias
- Update typed_container.h to redirect to simd_batch.h with warning
- Update core/container.h to include simd_batch.h instead
- Update documentation (API_REFERENCE, PROJECT_STRUCTURE, FEATURES)

Part of Issue #320, Closes #328
